Under python3 it is necessary to explicitly typecast the result of some divisions to integer types.  Patch extracted from series by Xavier Forestier: https://sourceforge.net/p/ffado/mailman/message/35457569/.
git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2659 2be59082-3212-0410-8809-b0798e1608f0

+++ support/mixer-qt4/ffado/mixer/audiofire.py
@@ -179,7 +179,7 @@ class AudioFire(QWidget):
         outputtabs = QTabWidget(self)
         outputtabslayout.addWidget( outputtabs, 1 )
 
-        for outpair in range(nb_pys_out/2):
+        for outpair in range(int(nb_pys_out/2)):
             tab = QWidget( outputtabs )
             tablayout = QHBoxLayout( tab )
             
@@ -200,7 +200,7 @@ class AudioFire(QWidget):
             grpMonitorLayout = QHBoxLayout()
             grpMonitor.setLayout(grpMonitorLayout);
             output_id = outpair * 2
-            for inpair in range(nb_pys_in/2):
+            for inpair in range(int(nb_pys_in/2)):
                 # create GUI elements
                 strip = AfMonitorWidget( grpMonitor )
                 grpMonitorLayout.addWidget( strip, 1 )
